I dunno what the curly braces mean in talonscript
I dunno what the curly braces mean in talonscript.  I kinda know the text is whatever was in the text variable from the line above.  So maybe the curly braces .are a way of sticking whatever was in the variable in the string.  I kinda think this is called string interpolation, and I kinda think that talonscript is based off of python.  When I mean based off of I mean it is a very limited subset of python, and that you can read about 'python f-strings' here: https://realpython.com/python-f-strings/  except these talon strings don't requir2e an f in front of them.