Learning to understand talon symbol syntax
This is the file where take everything I highlighted as condfusing and learn what it means.

I am confused about a bunch of things, which can be seen by reverting to this commit, opening up the webiste, and see what is highlighted in red.  Now I start the process of learning about the things highlighted in red that I do not know.

https://github.com/tararoys/LearningOverTime/commit/3b651f5eb640331adc1c261b5d873670c4f90fdd