Kinda know how to declare optional keywords in talon
The part that I know is that if you put square brackets around a word, that means that word is optional.

The part that I don't know is that there seems to be ongoing debate about whether these optional things are even a good idea.  I do not understand captures yet, but from the snippet I included in the long version it appears that allowing optional commands alows you to get spammed with a command running in the backgorund all the time.