I dunno what a captured phrase is
I dunno what a captured phrase is.  It looks like it is related to the <phrase> thing two commits ago that I am confused about.  I speculate that the <phrase> is part of a command that you can save and give to the right side stuff and use the right side stuff to useful thigns to the saved command.  Like the left side is the order, and the right side are the people carrying it out, and <phrase> is a way of remembering a useful peice of the order so the right side can do it's job.  But I dunno.