dunno what any of this is
I dunno anything about what is going on in this line.  I dunno what null coalescing is. I dunno what it means for a phrase to be optional.  (I kinda know that it has to do with the parts of the phrase that are in square brackets on the left hand side.  but I dunno.)