dunno what is  user.long_action_name(phrase, 1, 2, 3, 4) # call a longer action with parameters
I dunno what a user is.  I dunno what a long action name is.  I know what parameters are, but I dunno what phrase is and I dunno the what the numbers mean.  I dunno if a longer action just refers to the fact that the name is long, or if that has something to do with the duration of the action.