dunno what @mod.action_class is
I dunno what the @ does.  kinda know that mod is the variable name we made in line three.  I dunno what the . is.  I kinda think it has something to do with what they call 'accessing a property'  I dunno what action_class is.  I dunno what kind of property it is.