I dunno what is    def tab_jump(number: int):
I dunno what a def is.  I kinda know tab_jump is a name.  I kinda know def is a function declaration.  I kinda know that what is in the parenthesis are arguments.  I dunno what number : int is.  I kinda think it means that there is one argument, it has to be a number, and the number has to be an integer.

I dunno what the colon at the end is.

Things I kinda know: I kinda know thag this is a function declaration.  I kinda know that there is a difference between a declaration and an implementation.  I kinda know that a declaration is like reserving a space to have a party, instead of the actual party itself- it's basically saying I'm reserving space in this place to put something there later.

I kinda know this has something to do with namespaces.  I dunno exactly what a namespace is.